Reset
Taking a step back and focusing on data structure. How can book data be organised in a way that is both clear and easily translated to GeoJSON? To start with I've tried storing book route details in CSV and converting them with a script, with the result working nicely in https://geojson.io. This way it's up and running and I can start to build a little library. Even if the data structure changes down the line having a consistent format ought to make any changes pretty straightforward. Next step: rendering book routes on my own globe.